Solution to Check Alphabet Case

+++ Check Alphabet Case@@6788dd63490637cfc44e1b92/code.c
@@ -4,9 +4,9 @@ int main() {
     
     char ch;
     scanf("%c",&n);
-    if (ch>=65 && ch<=90)
+    if ((int)ch>=65 && (int)ch<=90)
     printf("Uppercase");
-    else if(ch>=97 && ch<=122)
+    else if((int)ch>=97 && (int)ch<=122)
     printf("Lowercase");
     return 0;
 }
